Refactor: [BADA-193] GifticonPage와 GifticonList의 데이터 처리 책임 분리#111
