+++ src/pages/trade/gifticon/page/GifticonPage.tsx
@@ -1,7 +1,8 @@
 'use client';
 
-import { useState } from 'react';
+import { useMemo, useState } from 'react';
 
+import { useTradePostsQuery } from '@/entities/trade-post/model/queries';
 import { GifticonFilter } from '@/pages/trade/gifticon/ui/GifticonFilter';
 import { GifticonList } from '@/pages/trade/gifticon/ui/GifticonList';
 import { useSortStateHook } from '@/shared/model/useSortStateHook';
@@ -12,12 +13,36 @@ import { TradeFloatingButton } from '@/widgets/trade/floating-button/ui/TradeFlo
 import { TradeSearchInput } from '@/widgets/trade/search-input/ui/TradeSearchInput';
 import { TradeSortFilter } from '@/widgets/trade/trade-sort-filter';
 
+const SORT_OPTIONS = [
+  { value: 'latest', label: '최신순' },
+  { value: 'popular', label: '인기순' },
+];
+
 export default function GifticonPage() {
   const [selectedCategory, setSelectedCategory] = useState<string>('전체');
   const { sortOption, setSortOption, isSortDrawerOpen, openDrawer, closeDrawer } = useSortStateHook<
     'latest' | 'popular'
   >('latest');
 
+  const { posts, isLoading } = useTradePostsQuery();
+
+  const processedPosts = useMemo(() => {
+    if (!posts) return [];
+
+    const gifticonPosts = posts.filter((p) => p.postCategory === 'GIFTICON');
+
+    const filtered = gifticonPosts.filter(
+      (p) => selectedCategory === '전체' || p.gifticonCategory === selectedCategory,
+    );
+
+    return [...filtered].sort((a, b) =>
+      sortOption === 'latest'
+        ? new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
+        : b.likesCount - a.likesCount,
+    );
+  }, [posts, selectedCategory, sortOption]);
+
+  const currentSortLabel = SORT_OPTIONS.find((o) => o.value === sortOption)?.label ?? '최신순';
   return (
     <>
       <BaseLayout
@@ -32,9 +57,10 @@ export default function GifticonPage() {
           onSelectCategory={setSelectedCategory}
         />
         <GifticonList
-          sortOption={sortOption}
+          items={processedPosts}
+          isLoading={isLoading}
+          sortLabel={currentSortLabel}
           onSortClick={openDrawer}
-          selectedCategory={selectedCategory}
         />
       </BaseLayout>
       <TradeSortFilter
